pyproject.toml
[tool.poetry]
name = "yandex-marks"
version = "0.1.0"
description = ""
authors = [" annavaleriev <filenko.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.5"
djangorestframework = "3.15.2"
psycopg2-binary = "2.9.9"
gunicorn = "21.2.0"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
